rename cliboa.client to cliboa.interface
