+++ cliboa/test/core/test_factory.py
@@ -16,7 +16,7 @@ import sys
 
 import pytest
 
-from cliboa.client import CommandArgumentParser
+from cliboa.interface import CommandArgumentParser
 from cliboa.core.factory import CustomInstanceFactory, ScenarioManagerFactory, StepExecutorFactory
 from cliboa.core.manager import JsonScenarioManager, YamlScenarioManager
 from cliboa.core.strategy import MultiProcExecutor, MultiProcWithConfigExecutor, SingleProcExecutor
