+++ cliboa/test/core/test_manager.py
@@ -20,7 +20,7 @@ from datetime import datetime, timedelta
 import pytest
 import yaml
 
-from cliboa.client import CommandArgumentParser
+from cliboa.interface import CommandArgumentParser
 from cliboa.conf import env
 from cliboa.core.manager import JsonScenarioManager, YamlScenarioManager
 from cliboa.core.scenario_queue import ScenarioQueue
