+++ cliboa/test/util/test_lisboa_log.py
@@ -15,7 +15,7 @@ import logging
 import os
 import sys
 
-from cliboa.client import CommandArgumentParser
+from cliboa.interface import CommandArgumentParser
 from cliboa.conf import env
 from cliboa.util.lisboa_log import LisboaLog
